@@ -15,10 +15,10 @@
  */
 package org.factcast.schema.registry.cli.commands
 
+import io.github.oshai.kotlinlogging.KotlinLogging
 import java.io.IOException
 import java.nio.file.Path
-import javax.inject.Singleton
-import mu.KotlinLogging
+import jakarta.inject.Singleton
 import org.factcast.schema.registry.cli.fs.FileSystemService
 import org.factcast.schema.registry.cli.project.ProjectService
 import org.factcast.schema.registry.cli.registry.DistributionCreatorService
@@ -85,15 +85,15 @@ class CommandServiceImpl(
                 logger.info("")
                 logger.error("Validation failed!")
 
-                 1
+                1
             }, {
                 logger.info("Project seems to be valid!")
 
-                 0
+                0
             })
     } catch (e: IllegalArgumentException) {
         logger.error(e) { "Invalid paths" }
 
-         1
+        1
     }
 }
