@@ -17,9 +17,9 @@ package org.factcast.schema.registry.cli.project.structure
 
 import io.micronaut.core.annotation.Introspected
 import java.nio.file.Path
-import javax.validation.Valid
-import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.NotNull
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotEmpty
+import jakarta.validation.constraints.NotNull
 import org.factcast.schema.registry.cli.domain.Event
 import org.factcast.schema.registry.cli.validation.NO_DESCRIPTION
 import org.factcast.schema.registry.cli.validation.NO_EVENT_VERSIONS
@@ -28,14 +28,14 @@ import org.factcast.schema.registry.cli.validation.NO_EVENT_VERSIONS
 data class EventFolder(
     override val path: Path,
 
-    @get:NotEmpty(message = NO_EVENT_VERSIONS)
-    @get:Valid
+    @field:NotEmpty(message = NO_EVENT_VERSIONS)
+    @field:Valid
     val versionFolders: List<EventVersionFolder>,
 
-    @get:NotNull(message = NO_DESCRIPTION)
+    @field:NotNull(message = NO_DESCRIPTION)
     val description: Path?,
 
-    @get:Valid
+    @field:Valid
     val transformationFolders: List<TransformationFolder>
 ) : Folder {
     override fun getChildren(): List<Folder> = versionFolders
