@@ -15,19 +15,19 @@
  */
 package org.factcast.schema.registry.cli.project.structure
 
+import io.github.oshai.kotlinlogging.KotlinLogging
 import io.micronaut.core.annotation.Introspected
+import jakarta.validation.Valid
 import java.nio.file.Path
-import javax.validation.Valid
-import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.NotNull
-import mu.KotlinLogging
+import jakarta.validation.constraints.NotEmpty
+import jakarta.validation.constraints.NotNull
 import org.factcast.schema.registry.cli.domain.Project
 import org.factcast.schema.registry.cli.validation.NO_DESCRIPTION
 import org.factcast.schema.registry.cli.validation.NO_NAMESPACES
 
 private val logger = KotlinLogging.logger {}
 
-@Introspected
+@Introspected(classes = [ProjectFolder::class, List::class])
 data class ProjectFolder(
     override val path: Path,
 
@@ -51,8 +51,8 @@ fun ProjectFolder.toProject() =
     )
 
 fun ProjectFolder.log() =
-        namespaces.flatMap { ns ->
-            ns.eventFolders.flatMap { folder ->
-                folder.versionFolders
-            }
-        }.forEach { logger.debug("Included event ${it.path}") }
+    namespaces.flatMap { ns ->
+        ns.eventFolders.flatMap { folder ->
+            folder.versionFolders
+        }
+    }.forEach { logger.debug("Included event ${it.path}") }
