@@ -22,8 +22,8 @@ import org.factcast.schema.registry.cli.project.structure.ProjectFolder
 import org.factcast.schema.registry.cli.project.structure.toProject
 import org.factcast.schema.registry.cli.validation.*
 import org.factcast.schema.registry.cli.validation.validators.ProjectStructureValidationService
-import javax.inject.Singleton
-import javax.validation.Validator
+import jakarta.inject.Singleton
+import jakarta.validation.Validator
 
 @Singleton
 class ProjectStructureValidationServiceImpl(private val validator: Validator) : ProjectStructureValidationService {
@@ -34,22 +34,30 @@ class ProjectStructureValidationServiceImpl(private val validator: Validator) : 
                 when (it.messageTemplate) {
                     NO_DESCRIPTION ->
                         ProjectError.NoDescription(it.leafBean.toFolder().path)
+
                     NO_EVENT_VERSIONS ->
                         ProjectError.NoEventVersions(it.leafBean.toFolder().path)
+
                     NO_SCHEMA ->
                         ProjectError.NoSchema(it.leafBean.toFolder().path)
+
                     NO_EXAMPLES ->
                         ProjectError.NoExamples(it.leafBean.toFolder().path)
+
                     NO_EVENTS ->
                         ProjectError.NoEvents(it.leafBean.toFolder().path)
+
                     NO_NAMESPACES ->
                         ProjectError.NoNamespaces(it.leafBean.toFolder().path)
+
                     NO_TRANSFORMATION_FILE ->
                         ProjectError.NoSuchFile(it.leafBean.toFolder().path)
+
                     TRANSFORMATION_VERSION_INVALID, VERSION_INVALID -> {
                         val folder = it.leafBean.toFolder()
                         ProjectError.WrongVersionFormat(folder.path.fileName.toString(), folder.path)
                     }
+
                     else ->
                         throw IllegalArgumentException("Unknown error type: ${it.messageTemplate}")
                 }
