@@ -26,7 +26,7 @@ import org.factcast.schema.registry.cli.validation.MissingTransformationCalculat
 import org.factcast.schema.registry.cli.validation.ProjectError
 import org.factcast.schema.registry.cli.validation.TransformationEvaluator
 import org.factcast.schema.registry.cli.validation.validators.TransformationValidationService
-import javax.inject.Singleton
+import jakarta.inject.Singleton
 
 @Singleton
 class TransformationValidationServiceImpl(
