@@ -20,7 +20,7 @@ import io.micronaut.validation.validator.constraints.ConstraintValidator
 import io.micronaut.validation.validator.constraints.ConstraintValidatorContext
 import org.factcast.schema.registry.cli.validation.validators.ValidTransformationFolder
 import java.nio.file.Path
-import javax.inject.Singleton
+import jakarta.inject.Singleton
 
 @Singleton
 class ValidTransformationFolderValidator : ConstraintValidator<ValidTransformationFolder, Path> {
