@@ -18,7 +18,7 @@ package org.factcast.schema.registry.cli.validation.validators.impl
 import io.micronaut.core.annotation.AnnotationValue
 import io.micronaut.validation.validator.constraints.ConstraintValidator
 import java.nio.file.Path
-import javax.inject.Singleton
+import jakarta.inject.Singleton
 import org.factcast.schema.registry.cli.validation.validators.ValidVersionFolder
 
 @Singleton
