@@ -77,7 +77,9 @@ public interface FactCast extends ReadFactCast {
     return lock(list);
   }
 
-  /** @deprecated use lock(FactSpec) instead */
+  /**
+   * @deprecated use lock(FactSpec) instead
+   */
   // @Deprecated(forRemoval = true)
   @Deprecated
   DeprecatedLockedOperationBuilder lock(@NonNull String ns);
