@@ -70,7 +70,9 @@ public class RedissonTxManager {
     return block.apply(currentTx);
   }
 
-  /** @return true if tx was started, false if there was one running */
+  /**
+   * @return true if tx was started, false if there was one running
+   */
   public boolean startOrJoin() {
     if (currentTx == null) {
       currentTx = redisson.createTransaction(options);
