@@ -20,17 +20,19 @@ import io.grpc.ClientInterceptor;
 import java.util.*;
 import javax.annotation.Nullable;
 import lombok.NonNull;
+import net.devh.boot.grpc.client.autoconfigure.*;
 import net.devh.boot.grpc.client.channelfactory.GrpcChannelConfigurer;
 import net.devh.boot.grpc.client.channelfactory.GrpcChannelFactory;
 import org.factcast.client.grpc.FactCastGrpcClientProperties;
 import org.factcast.client.grpc.GrpcFactStore;
 import org.factcast.core.store.FactStore;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Lazy;
 
@@ -40,9 +42,19 @@ import org.springframework.context.annotation.Lazy;
  * @author uwe.schaefer@prisma-capacity.eu
  */
 @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
-@Configuration
+@AutoConfiguration
 @ConditionalOnClass({GrpcFactStore.class, GrpcChannelFactory.class})
 @Import(FactCastGrpcClientProperties.class)
+// spring-grpc compat until release
+@ImportAutoConfiguration({
+  GrpcClientAutoConfiguration.class,
+  GrpcClientMetricAutoConfiguration.class,
+  GrpcClientHealthAutoConfiguration.class,
+  GrpcClientSecurityAutoConfiguration.class,
+  GrpcClientTraceAutoConfiguration.class,
+  GrpcDiscoveryClientAutoConfiguration.class
+})
+// spring-grpc compat until release
 @EnableConfigurationProperties
 public class GrpcFactStoreAutoConfiguration {
 
