@@ -19,13 +19,13 @@ import lombok.Generated;
 import net.jpountz.lz4.LZ4Compressor;
 import org.factcast.server.grpc.FactStoreGrpcService;
 import org.factcast.server.grpc.codec.Lz4GrpcServerCodec;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
 @Generated
-@Configuration
+@AutoConfiguration
 @ConditionalOnClass({FactStoreGrpcService.class, LZ4Compressor.class, Lz4GrpcServerCodec.class})
 public class LZ4ServerAutoConfiguration {
   @Bean
