@@ -18,16 +18,16 @@ package org.factcast.spring.boot.autoconfigure.server.grpc;
 import lombok.Generated;
 import org.factcast.server.grpc.FactStoreGrpcService;
 import org.factcast.server.grpc.codec.SnappyGrpcServerCodec;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.xerial.snappy.SnappyInputStream;
 import org.xerial.snappy.SnappyOutputStream;
 
 @Generated
-@Configuration
+@AutoConfiguration
 @ConditionalOnClass({
   FactStoreGrpcService.class,
   SnappyInputStream.class,
