@@ -16,12 +16,12 @@
 package org.factcast.spring.boot.autoconfigure.store;
 
 import org.factcast.store.PgFactStoreConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
-@Configuration
+@AutoConfiguration
 @EnableConfigurationProperties
 @ConditionalOnClass(PgFactStoreConfiguration.class)
 @Import(PgFactStoreConfiguration.class)
