@@ -17,12 +17,12 @@ package org.factcast.spring.boot.autoconfigure.store.script.engine;
 
 import org.factcast.store.internal.script.JSEngineFactory;
 import org.factcast.store.internal.script.graaljs.GraalJSEngineFactory;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
+@AutoConfiguration
 @ConditionalOnClass({JSEngineFactory.class, GraalJSEngineFactory.class})
 public class GraalJSScriptEngineAutoConfiguration {
 
