@@ -33,7 +33,9 @@ public class PgLatestSerialFetcher {
 
   @NonNull final JdbcTemplate jdbcTemplate;
 
-  /** @return 0, if no Fact is found, or exception is raised. */
+  /**
+   * @return 0, if no Fact is found, or exception is raised.
+   */
   public long retrieveLatestSer() {
     SqlRowSet rs = jdbcTemplate.queryForRowSet(PgConstants.LAST_SERIAL_IN_LOG);
     if (rs.next()) {
