@@ -19,10 +19,10 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 import io.micrometer.core.instrument.Tags;
+import jakarta.servlet.http.HttpServletResponse;
 import java.net.URL;
 import java.util.Date;
 import java.util.Map;
-import javax.servlet.http.HttpServletResponse;
 import lombok.SneakyThrows;
 import okhttp3.*;
 import okhttp3.Response.Builder;
@@ -49,14 +49,15 @@ public class HttpIndexFetcherTest {
             String sinceHeader = headers.get(ValidationConstants.HTTPHEADER_IF_MODIFIED_SINCE);
 
             if (etag.equals(etagHeader) && since.equals(sinceHeader)) {
-              ctx.res.setStatus(304);
+              ctx.res().setStatus(304);
             } else {
-              HttpServletResponse res = ctx.res;
+              HttpServletResponse res = ctx.res();
               res.setStatus(200);
-              res.getWriter()
+              res.getOutputStream()
                   .write(
-                      "\n"
-                          + "{\"schemes\":[{\"id\":\"namespaceA/eventA/1/schema.json\",\"ns\":\"namespaceA\",\"type\":\"eventA\",\"version\":1,\"hash\":\"84e69a2d3e3d195abb986aad22b95ffd\"},{\"id\":\"namespaceA/eventA/2/schema.json\",\"ns\":\"namespaceA\",\"type\":\"eventA\",\"version\":2,\"hash\":\"24d48268356e3cb7ac2f148850e4aac1\"}]}");
+                      ("\n"
+                              + "{\"schemes\":[{\"id\":\"namespaceA/eventA/1/schema.json\",\"ns\":\"namespaceA\",\"type\":\"eventA\",\"version\":1,\"hash\":\"84e69a2d3e3d195abb986aad22b95ffd\"},{\"id\":\"namespaceA/eventA/2/schema.json\",\"ns\":\"namespaceA\",\"type\":\"eventA\",\"version\":2,\"hash\":\"24d48268356e3cb7ac2f148850e4aac1\"}]}")
+                          .getBytes());
               res.addHeader(ValidationConstants.HTTPHEADER_E_TAG, etag);
               res.addHeader(ValidationConstants.HTTPHEADER_LAST_MODIFIED, since);
             }
@@ -74,7 +75,7 @@ public class HttpIndexFetcherTest {
   void testThrowsExceptionOn404() throws Exception {
     try (TestHttpServer s = new TestHttpServer()) {
 
-      s.get("/registry/index.json", ctx -> ctx.res.setStatus(404));
+      s.get("/registry/index.json", ctx -> ctx.res().setStatus(404));
 
       var registryMetrics = mock(RegistryMetrics.class);
 
